Finishing everything the previous commit said it would do.
