+++ GPS.Service/Pages/Index.cshtml.cs
