+++ GPS.Service/Program.cs
@@ -1,28 +1,26 @@
-﻿using GPS.Data;
-using System;
-using System.Collections.Generic;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
-namespace GPS.Service
+namespace GPS.WebApp
 {
-    class Program
+    public class Program
     {
-        static void Main(string[] args)
+        public static void Main(string[] args)
         {
-            //Get a list of all outages from the .sof file
-            Parser sofParser = new Parser("SOF\\current.sof");
-            List<Outage> allOutages = sofParser.PopulateObjectsFromSof();
-
-            //Example of LINQ
-            foreach (var outages in
-            from Outage outages in allOutages
-            where outages.StartYear == 2020
-            select outages)
-            {
-                Console.WriteLine(outages.StartYear);
-            }
+            CreateHostBuilder(args).Build().Run();
         }
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
