+++ GPS.Service/wwwroot/lib/bootstrap/LICENSE
