+++ GPS.Service/wwwroot/lib/jquery-validation/LICENSE.md
