+++ GPS.Service/wwwroot/lib/jquery/LICENSE.txt
